Squashed from merge-vibe-vibe-20260522 at 193afce835bfdc408364bdf5b8795e10102780ea.

This integrates the accumulated upstream Sunshine changes into Apollo as a single local history point.

+++ src/display_device.cpp
@@ -35,7 +35,7 @@ namespace display_device {
       if (result > std::numeric_limits<unsigned int>::max()) {
         throw std::out_of_range("stou");
       }
-      return result;
+      return (int) result;
     }
 
     /**
